==> 01_data_cleaning.ipynb <==
import geopandas as gpd
import pandas as pd
import datetime

incendis_86_20 = gpd.read_file("dades/incendis_86_20.geojson")

# Comprovo si hi han duplicats
incendis_86_20.geometry.duplicated().sum()

incendis_86_20.drop_duplicates(subset=['geometry'], inplace=True)

# Comprovació dels nuls en la data
incendis_86_20[incendis_86_20["DATA_INCEN"].isna()]

incendis_86_20.dropna(subset=["DATA_INCEN"], inplace=True)

incendis_86_20["DATA_INCEN"]

Hi han dates amb l'any en format de 2 xifres hi altres amb format de 4 xifres, a part d'algun format erroni.

# Estandarització de formats
incendis_86_20["DATA_INCEN"] = incendis_86_20["DATA_INCEN"].str.split('\r').str[0]
date1 = pd.to_datetime(incendis_86_20["DATA_INCEN"], errors='coerce', format='%d/%m/%y')
date2 = pd.to_datetime(incendis_86_20["DATA_INCEN"], errors='coerce', format='%d/%m/%Y')
incendis_86_20["date"] = date1.fillna(date2)
incendis_86_20.dropna(subset=["date"], inplace=True)
incendis_86_20["year"] = incendis_86_20["date"].dt.year

# Exporto
incendis_86_20.to_file('dades/incendis_86_20_clean.geojson', driver='GeoJSON')
