==> 09_biom_carbon_quant.ipynb <==
import pandas as pd
import numpy as np

# Incendi Baldomà

zonal_stats_df = pd.read_csv("dades/resultats_baldoma/stats_arbrat.csv")

### Coníferes

cnf = zonal_stats_df[zonal_stats_df["coberta_sol"]== 2]
# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_con = {1:0.25, 2:0.47, 3:0.56, 4:0.65}
# Biomassa Consumida (factor)
cnf["biomassa_cons"] = cnf["Severitat"].map(dict_sv_bm_con)
# Biomassa preexistent
cnf["biomassa_pre"] = cnf["bm_mean"] * cnf["area"]
# Biomassa consumida (tonelades)
cnf["biomassa_con_tot"] = cnf["biomassa_pre"] * cnf["biomassa_cons"]
# CO2 allibertat
cnf["CO2"] = cnf["biomassa_con_tot"] * 1.377
# CH4 allibertat
cnf["CH4"] = cnf["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
cnf["CO2eq_CH4"] = cnf["CH4"] * 21
# N2O allibertat
cnf["N2O"] = cnf["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
cnf["CO2eq_N2O"] = cnf["N2O"] * 298
#b CO2 equivalent total 
cnf["CO2eq_total"] = cnf["CO2"] + cnf["CO2eq_CH4"] + cnf["CO2eq_N2O"]
# Exporta
cnf.to_excel("dades/resultats_baldoma/con_baldoma.xlsx")

### Frondoses

frnd = zonal_stats_df[zonal_stats_df["coberta_sol"].isin([3, 4, 7])]
# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_con = {1:0.25, 2:0.4, 3:0.48, 4:0.56}
# Biomassa Consumida (factor)
frnd["biomassa_cons"] = frnd["Severitat"].map(dict_sv_bm_con)
# Biomassa preexistent
frnd["biomassa_pre"] = frnd["bm_mean"] * frnd["area"]
# Biomassa consumida (tonelades)
frnd["biomassa_con_tot"] = frnd["biomassa_pre"] * frnd["biomassa_cons"]
frnd = frnd.groupby(["Severitat"])[["area", "biomassa_pre", "biomassa_con_tot"]].sum()
# CO2 allibertat
frnd["CO2"] = frnd["biomassa_con_tot"] * 1.377
# CH4 allibertat
frnd["CH4"] = frnd["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
frnd["CO2eq_CH4"] = frnd["CH4"] * 21
# N2O allibertat
frnd["N2O"] = frnd["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
frnd["CO2eq_N2O"] = frnd["N2O"] * 298
#b CO2 equivalent total 
frnd["CO2eq_total"] = frnd["CO2"] + frnd["CO2eq_CH4"] + frnd["CO2eq_N2O"]
# Exporta
frnd.to_excel("dades/resultats_baldoma/frnd_baldoma.xlsx")

### Estrat arbustiu

arbustiu = pd.read_csv("dades/resultats_baldoma/stats_arbustiu.csv")

# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_arb = {1:0.71, 2:0.84, 3:0.89, 4:0.95}
# Biomassa garrics
arbustiu.loc[arbustiu["n_FORmat"]==160, "biomassa"] = np.e**(-2.892 + 1.505 * np.log(arbustiu["HMMAT"]) + 0.462 * np.log(arbustiu["FCCMAT"]))
# Biomassa boixedes, Tanques, orles, bardes, salcedes, galeries arbustives, etc
arbustiu.loc[arbustiu["n_FORmat"].isin([120, 110, 111]), "biomassa"] = np.e**(0.949 * np.log(arbustiu["FCCMAT"]))
# Biomassa Consumida (factor)
arbustiu["biomassa_cons"] = arbustiu["Severitat"].map(dict_sv_bm_arb)
# Biomassa preexistent
arbustiu["biomassa_pre"] = arbustiu["biomassa"] * arbustiu["Superficie"]
# Biomassa consumida (tonelades)
arbustiu["biomassa_con_tot"] = arbustiu["biomassa_pre"] * arbustiu["biomassa_cons"]
arbustiu = arbustiu.groupby(["Severitat"])[["Superficie", "biomassa_pre", "biomassa_con_tot"]].sum()
# CO2 allibertat
arbustiu["CO2"] = arbustiu["biomassa_con_tot"] * 1.377
# CH4 allibertat
arbustiu["CH4"] = arbustiu["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
arbustiu["CO2eq_CH4"] = arbustiu["CH4"] * 21
# N2O allibertat
arbustiu["N2O"] = arbustiu["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
arbustiu["CO2eq_N2O"] = arbustiu["N2O"] * 298
#b CO2 equivalent total 
arbustiu["CO2eq_total"] = arbustiu["CO2"] + arbustiu["CO2eq_CH4"] + arbustiu["CO2eq_N2O"]
# Exporta
arbustiu.to_excel("dades/resultats_baldoma/arbustiu_baldoma.xlsx")

# Incendi El Pont de Vilomara

zonal_stats_df = pd.read_csv("dades/resultats_pont_vilomara/stats_arbrat.csv")
zonal_stats_df.sort_values("Severitat", inplace=True)

### Coníferes

cnf = zonal_stats_df[zonal_stats_df["tipus"]== 2]
# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_con = {1:0.25, 2:0.47, 3:0.56, 4:0.65}
# Biomassa Consumida (factor)
cnf["biomassa_cons"] = cnf["Severitat"].map(dict_sv_bm_con)
# Biomassa preexistent
cnf["biomassa_pre"] = cnf["bm_mean"] * cnf["area"]
# Biomassa consumida (tonelades)
cnf["biomassa_con_tot"] = cnf["biomassa_pre"] * cnf["biomassa_cons"]
# CO2 allibertat
cnf["CO2"] = cnf["biomassa_con_tot"] * 1.377
# CH4 allibertat
cnf["CH4"] = cnf["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
cnf["CO2eq_CH4"] = cnf["CH4"] * 21
# N2O allibertat
cnf["N2O"] = cnf["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
cnf["CO2eq_N2O"] = cnf["N2O"] * 298
# CO2 equivalent total 
cnf["CO2eq_total"] = cnf["CO2"] + cnf["CO2eq_CH4"] + cnf["CO2eq_N2O"]
# Exporta
cnf.to_excel("dades/resultats_pont_vilomara/con_pont_vilomara.xlsx")

### Frondoses

frnd = zonal_stats_df[zonal_stats_df["tipus"].isin([3, 4, 7])]
# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_con = {1:0.25, 2:0.4, 3:0.48, 4:0.56}
# Biomassa Consumida (factor)
frnd["biomassa_cons"] = frnd["Severitat"].map(dict_sv_bm_con)
# Biomassa preexistent
frnd["biomassa_pre"] = frnd["bm_mean"] * frnd["area"]
# Biomassa consumida (tonelades)
frnd["biomassa_con_tot"] = frnd["biomassa_pre"] * frnd["biomassa_cons"]
frnd = frnd.groupby(["Severitat"])[["area", "biomassa_pre", "biomassa_con_tot"]].sum()
# CO2 allibertat
frnd["CO2"] = frnd["biomassa_con_tot"] * 1.377
# CH4 allibertat
frnd["CH4"] = frnd["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
frnd["CO2eq_CH4"] = frnd["CH4"] * 21
# N2O allibertat
frnd["N2O"] = frnd["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
frnd["CO2eq_N2O"] = frnd["N2O"] * 298
# CO2 equivalent total 
frnd["CO2eq_total"] = frnd["CO2"] + frnd["CO2eq_CH4"] + frnd["CO2eq_N2O"]
# Exporta
frnd.to_excel("resultats_pont_vilomara/frnd_pont_vilomara.xlsx")

### Estrat arbustiu

arbustiu = pd.read_csv("resultats_pont_vilomara/stats_arbustiu.csv")
# Diccionari amb el mapeix severitat - Factor de biomassa consumida 
dict_sv_bm_arb = {1:0.71, 2:0.84, 3:0.89, 4:0.95}
# Biomassa garrics
arbustiu.loc[arbustiu["n_FORmat"]==160, "biomassa"] = np.e**(-2.892 + 1.505 * np.log(arbustiu["HMMAT"]) + 0.462 * np.log(arbustiu["FCCMAT"]))
# Biomassa boixedes, Tanques, orles, bardes, salcedes, galeries arbustives, etc
arbustiu.loc[arbustiu["n_FORmat"].isin([120, 110, 113]), "biomassa"] = np.e**(0.949 * np.log(arbustiu["FCCMAT"]))
# Biomassa timonedes
arbustiu.loc[arbustiu["n_FORmat"]==250, "biomassa"] =  np.e**(-1.887 + 0.643 * np.log(arbustiu["HMMAT"]) + 0.661 * np.log(arbustiu["FCCMAT"]))
# Biomassa escobillars
arbustiu.loc[arbustiu["n_FORmat"]==255, "biomassa"] =  np.e**(-1.833 + 0.965 * np.log(arbustiu["HMMAT"]) + 0.453 * np.log(arbustiu["FCCMAT"]))
# Biomassa Consumida (factor)
arbustiu["biomassa_cons"] = arbustiu["Severitat"].map(dict_sv_bm_arb)
# Biomassa preexistent
arbustiu["biomassa_pre"] = arbustiu["biomassa"] * arbustiu["superficie"]
# Biomassa consumida (tonelades)
arbustiu["biomassa_con_tot"] = arbustiu["biomassa_pre"] * arbustiu["biomassa_cons"]
arbustiu = arbustiu.groupby(["Severitat"])[["superficie", "biomassa_pre", "biomassa_con_tot"]].sum()
# CO2 allibertat
arbustiu["CO2"] = arbustiu["biomassa_con_tot"] * 1.377
# CH4 allibertat
arbustiu["CH4"] = arbustiu["biomassa_con_tot"] * 0.0014
# CO2 equivalent corresponent al CH4
arbustiu["CO2eq_CH4"] = arbustiu["CH4"] * 21
# N2O allibertat
arbustiu["N2O"] = arbustiu["biomassa_con_tot"] * 0.00012
# CO2 equivalent corresponent al N2O
arbustiu["CO2eq_N2O"] = arbustiu["N2O"] * 298
#b CO2 equivalent total 
arbustiu["CO2eq_total"] = arbustiu["CO2"] + arbustiu["CO2eq_CH4"] + arbustiu["CO2eq_N2O"]
# Exporta
arbustiu.to_excel("resultats_pont_vilomara/arbustiu_pont_vilomara.xlsx")
